+++ inference-engine/src/mkldnn_plugin/nodes/mkldnn_concat_node.cpp
@@ -17,7 +17,7 @@
 #include "mkldnn_memory.h"
 #include "ie_parallel.hpp"
 #include "mkldnn_conv_node.h"
-#include "mkldnn_quantize_node.h"
+#include "mkldnn_fake_quantize_node.h"
 #include "mkldnn_pooling_node.h"
 #include "mkldnn_eltwise_node.h"
 #include <limits>
