+++ inference-engine/src/mkldnn_plugin/nodes/mkldnn_conv_node.cpp
@@ -6,7 +6,7 @@
 #include "mkldnn_reorder_node.h"
 #include "mkldnn_input_node.h"
 #include "mkldnn_eltwise_node.h"
-#include "mkldnn_quantize_node.h"
+#include "mkldnn_fake_quantize_node.h"
 #include "mkldnn_pooling_node.h"
 #include "mkldnn_concat_node.h"
 #include <string>
