+++ inference-engine/src/mkldnn_plugin/nodes/mkldnn_fullyconnected_node.cpp
@@ -4,7 +4,7 @@
 
 #include "mkldnn_fullyconnected_node.h"
 #include "mkldnn_eltwise_node.h"
-#include "mkldnn_quantize_node.h"
+#include "mkldnn_fake_quantize_node.h"
 
 #include <legacy/ie_layers.h>
 #include <string>
